+++ src/main/java/org/recordy/server/common/message/ErrorMessage.java
@@ -1,0 +1,14 @@
+package org.recordy.server.common.message;
+
+import lombok.Getter;
+import lombok.RequiredArgsConstructor;
+import org.springframework.http.HttpStatus;
+
+@RequiredArgsConstructor
+@Getter
+public enum ErrorMessage {
+
+    ;
+    private final HttpStatus httpStatus;
+    private final String message;
+}
